+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 3.1.0
 ### tmux
 - add tmux 3.2a to Travis test matrix; add 3.2a to supported tmux versions list
 ### Misc